requirements.txt
requests
icalendar
pytz
PyYAML
python-dotenv
# Testing dependencies  
pytest
pytest-cov